Unmiddleware the appconfig validation
