+++ server/service/appconfig.go
@@ -152,6 +152,19 @@ func (svc *Service) ModifyAppConfig(ctx context.Context, p []byte) (*fleet.AppCo
 		return nil, err
 	}
 
+	// TODO(mna): this ports the validations from the old validationMiddleware
+	// correctly, but this could be optimized so that we don't unmarshal the
+	// incoming bytes twice.
+	invalid := &fleet.InvalidArgumentError{}
+	var newAppConfig fleet.AppConfig
+	if err := json.Unmarshal(p, &newAppConfig); err != nil {
+		return nil, ctxerr.Wrap(ctx, err)
+	}
+	validateSSOSettings(newAppConfig, appConfig, invalid)
+	if invalid.HasErrors() {
+		return nil, ctxerr.Wrap(ctx, invalid)
+	}
+
 	// We apply the config that is incoming to the old one
 	decoder := json.NewDecoder(bytes.NewReader(p))
 	decoder.DisallowUnknownFields()
@@ -172,6 +185,37 @@ func (svc *Service) ModifyAppConfig(ctx context.Context, p []byte) (*fleet.AppCo
 		return nil, err
 	}
 	return appConfig, nil
+}
+
+func validateSSOSettings(p fleet.AppConfig, existing *fleet.AppConfig, invalid *fleet.InvalidArgumentError) {
+	if p.SSOSettings.EnableSSO {
+		if p.SSOSettings.Metadata == "" && p.SSOSettings.MetadataURL == "" {
+			if existing.SSOSettings.Metadata == "" && existing.SSOSettings.MetadataURL == "" {
+				invalid.Append("metadata", "either metadata or metadata_url must be defined")
+			}
+		}
+		if p.SSOSettings.Metadata != "" && p.SSOSettings.MetadataURL != "" {
+			invalid.Append("metadata", "both metadata and metadata_url are defined, only one is allowed")
+		}
+		if p.SSOSettings.EntityID == "" {
+			if existing.SSOSettings.EntityID == "" {
+				invalid.Append("entity_id", "required")
+			}
+		} else {
+			if len(p.SSOSettings.EntityID) < 5 {
+				invalid.Append("entity_id", "must be 5 or more characters")
+			}
+		}
+		if p.SSOSettings.IDPName == "" {
+			if existing.SSOSettings.IDPName == "" {
+				invalid.Append("idp_name", "required")
+			}
+		} else {
+			if len(p.SSOSettings.IDPName) < 4 {
+				invalid.Append("idp_name", "must be 4 or more characters")
+			}
+		}
+	}
 }
 
 ////////////////////////////////////////////////////////////////////////////////
