+++ server/service/appconfig_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/fleetdm/fleet/v4/server/fleet"
 	"github.com/fleetdm/fleet/v4/server/mock"
 	"github.com/fleetdm/fleet/v4/server/ptr"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -219,4 +220,43 @@ func echoHandler() http.Handler {
 func setupCertificateChain(t *testing.T) (server *httptest.Server, teardown func()) {
 	server = httptest.NewUnstartedServer(echoHandler())
 	return server, server.Close
+}
+
+func TestSSONotPresent(t *testing.T) {
+	invalid := &fleet.InvalidArgumentError{}
+	var p fleet.AppConfig
+	validateSSOSettings(p, &fleet.AppConfig{}, invalid)
+	assert.False(t, invalid.HasErrors())
+
+}
+
+func TestNeedFieldsPresent(t *testing.T) {
+	invalid := &fleet.InvalidArgumentError{}
+	config := fleet.AppConfig{
+		SSOSettings: fleet.SSOSettings{
+			EnableSSO:   true,
+			EntityID:    "fleet",
+			IssuerURI:   "http://issuer.idp.com",
+			MetadataURL: "http://isser.metadata.com",
+			IDPName:     "onelogin",
+		},
+	}
+	validateSSOSettings(config, &fleet.AppConfig{}, invalid)
+	assert.False(t, invalid.HasErrors())
+}
+
+func TestMissingMetadata(t *testing.T) {
+	invalid := &fleet.InvalidArgumentError{}
+	config := fleet.AppConfig{
+		SSOSettings: fleet.SSOSettings{
+			EnableSSO: true,
+			EntityID:  "fleet",
+			IssuerURI: "http://issuer.idp.com",
+			IDPName:   "onelogin",
+		},
+	}
+	validateSSOSettings(config, &fleet.AppConfig{}, invalid)
+	require.True(t, invalid.HasErrors())
+	assert.Contains(t, invalid.Error(), "metadata")
+	assert.Contains(t, invalid.Error(), "either metadata or metadata_url must be defined")
 }
